+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2023/06/23
+### Added
+- Added support for XML string/tuning manifests.
+### Changed
+- SimData files are now written with the `.SimData.xml` extension.
+- Manifests now enforce "properties", "json", or "xml" types.
+### Fixed
+- Fixed issue where strings whose keys start with "0x0" would not be mapped.
+
 ## [0.2.0] - 2023/05/23
 ### Added
 - Added support for indexing SDX packages (must add folders to srcDirs though).